+++ src/test/java/com/codeit/deokhugam/service/PatchReviewServiceTest.java
@@ -59,7 +59,7 @@ public class PatchReviewServiceTest {
   String newContent;
   PatchReviewCommand patchReviewCommand;
   LocalDate publishDate;
-  int reviewCount;
+  long reviewCount;
   String isbn;
   String thumbnailUrl;
   BigDecimal ratingInBook;
@@ -78,7 +78,7 @@ public class PatchReviewServiceTest {
     description = "description";
     publisher = "publisher";
     publishDate = LocalDate.now();
-    reviewCount = 0;
+    reviewCount = 0L;
     ratingInBook = BigDecimal.valueOf(1.1);
     isbn = "isbn";
     thumbnailUrl = "thumbnailUrl";
